Add splash screen and animated logo widget
Added a new splash screen with an animated logo widget to the project. The `AnimatedLogo` widget is responsible for animating the logo on the splash screen. The `SplashViewBody` widget is responsible for displaying the logo and starting the animation. The `SplashView` widget is the main widget that displays the `SplashViewBody` widget.