Refactor custom app bar and add custom app bar color
The commit refactors the custom app bar widget by importing a color from the ColorManager class and using it as the background color for the app bar. Additionally, a new custom app bar color is added to the ColorManager class. A new text style is also added to the Styles class.